requirements.txt
fastapi
python-dotenv
uvicorn
python-multipart
protobuf==3.20.1
flake8
black
flake8-docstrings
pep8-naming
isort
